+++ Models/Sampling/ISampleCollection.cs
@@ -13,7 +13,7 @@ public interface ISampleCollection
     /// </summary>
     /// <param name="x">Sample x-coordinate.</param>
     /// <param name="y">Sample y-coordinate.</param>
-    public void Add(double x, double y);
+    public void Add(float x, float y);
 
     /// <summary>
     /// Returns an enumerable of neighbour samples within the specified distance from the given coordinates.
@@ -22,5 +22,5 @@ public interface ISampleCollection
     /// <param name="y">Sample y-coordinate.</param>
     /// <param name="distance">The distance within which to search for neighbour samples. How this distance is measured depends on the implementation.</param>
     /// <returns>An enumerable of neighbour samples.</returns>
-    IEnumerable<Neighbour> Neighbours(double x, double y, double distance);
+    IEnumerable<Neighbour> Neighbours(float x, float y, float distance);
 }
